==> 01_images.Rmd <==
---
title: "Images"
output:
  rmarkdown::github_document:
    df_print: kable
bibliography: ref-mono.bib
csl: associacao-brasileira-de-normas-tecnicas-ipea.csl
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook for the development and saving of the images used in the undergrad thesis. Some images will be developed on my own, some will be replications of cited papers. The notebook will follow the structure of the undergrad thesis.

```{r}
library("dotenv")
library("scales")
library("tidyverse")
library("duckdb")
library("gtsummary")
library("labelled")
library("ggthemes")
library("knitr")
```

```{r}
setwd("D:/OneDrive/R Workspace/computers_and_jobs")
load_dot_env()
base_dir <- "D:/OneDrive/R Workspace/computers_and_jobs"
fig_dir <- str_interp("${base_dir}/fig")
```

# Introduction

The first image used in the introduction section is a reproduction of @autor2019.

```{r include=FALSE}
source("./Replication data for autor 2019/do/occ_changes_plots_adjusted.R")

# Main Function
df <- as.data.table(read.dta13(data_path))
```

**Changes in Occupational Employment Shares, 1970-2016**

*Working Age Adults (Percent Change Over Decade)*

```{r warning=FALSE}
# Figure 3
plot_occ_group_pct_changes(df,subgroup='overall',subgroup_title='Working Age Adults',
										intervals=c('1970-1980','1980-1990','1990-2000','2000-2016'))
```

# Descriptive statistics

## RAIS and skill classification

First, I'll describe the outcome variables (mean wages and number of jobs) in general and segmented by skill type of the occupation, both for the first period and on average for all periods.

```{r}
datadir <- Sys.getenv("DATA_DIR")
file <- str_interp('${datadir}Final_db.parquet')
temp_db_dir <- Sys.getenv("DUCK_DB_DIR")
temp_db_mem <- Sys.getenv("DUCK_DB_MEM")

con <- dbConnect(duckdb(), dbdir = temp_db_dir)
dbExecute(conn = con, str_interp("PRAGMA memory_limit='${temp_db_mem}'"))
```

```{r}
tictoc::tic()

data <- tbl(con, file) %>%
  filter(ano < 2020) %>%
  group_by(
    ano,
    group
  ) %>%
  summarise(
    valor_remuneracao_media=mean(valor_remuneracao_media),
    valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
    mean_wage=mean(mean_wage),
    job_number=sum(job_number)
  ) %>%
  ungroup() %>%
  mutate(
    log_wage = log(mean_wage + 1)
  ) %>%
  collect()

data <- data %>%
  mutate(
    group = as.factor(group) %>% relevel('NRM'),
    ano = as.factor(ano)
  )

tictoc::toc()
data %>% head()
```

```{r}
data %>%
  filter(ano==2006) %>%
  group_by(
    group
  ) %>%
  summarise(
    valor_remuneracao_media=mean(valor_remuneracao_media),
    valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
    mean_wage=mean(mean_wage),
    job_number=sum(job_number)
  ) %>%
  mutate(
    job_share=job_number/sum(job_number)
  ) %>%
  ungroup() %>%
  mutate(
    mean_wage=label_dollar(prefix = 'R$ ')(mean_wage),
    job_number=label_number(big.mark = ',')(job_number),
    job_share=label_percent(accuracy=.01)(job_share)
  ) %>%
  select(group, mean_wage, job_number, job_share) %>%
  kable(
    col.names = c(
      'Skill group',
      'Mean hourly wage',
      'Number of emp. relationships',
      'Share of emp. relationships'
    ),
    caption = "Summary of wages and employment relationships for skill groups, 2006",
  align = "llrr"
  )
```

```{r}
data %>%
  filter(ano==2006) %>%
  summarise(
    valor_remuneracao_media=mean(valor_remuneracao_media),
    valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
    mean_wage=mean(mean_wage),
    job_number=sum(job_number)
  ) %>%
  mutate(
    group='All groups',
    job_share=job_number/sum(job_number),
    mean_wage=label_dollar(prefix = 'R$ ')(mean_wage),
    job_number=label_number(big.mark = ',')(job_number),
    job_share=label_percent(accuracy=.01)(job_share)
  ) %>%
  select(group, mean_wage, job_number, job_share) %>%
  kable(
    col.names = c(
      'Skill group',
      'Mean hourly wage',
      'Number of emp. relationships',
      'Share of emp. relationships'
    ),
    caption = "Summary of wages and employment relationships, 2006",
  align = "llrr"
  )
```

```{r}
data %>%
  filter(ano==2019) %>%
  group_by(
    group
  ) %>%
  summarise(
    valor_remuneracao_media=mean(valor_remuneracao_media),
    valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
    mean_wage=mean(mean_wage),
    job_number=sum(job_number)
  ) %>%
  mutate(
    job_share=job_number/sum(job_number)
  ) %>%
  ungroup() %>%
  mutate(
    mean_wage=label_dollar(prefix = 'R$ ')(mean_wage),
    job_number=label_number(big.mark=',')(job_number),
    job_share=label_percent(accuracy=.01)(job_share)
  ) %>%
  select(group, mean_wage, job_number, job_share) %>%
  kable(
    col.names = c(
      'Skill group',
      'Mean hourly wage',
      'Number of emp. relationships',
      'Share of emp. relationships'
    ),
    caption = "Summary of wages and employment relationships for skill groups, 2019",
  align = "llrr"
  )
```

```{r}
data %>%
  filter(ano==2019) %>%
  summarise(
    valor_remuneracao_media=mean(valor_remuneracao_media),
    valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
    mean_wage=mean(mean_wage),
    job_number=sum(job_number)
  ) %>%
  mutate(
    group='All groups',
    job_share=job_number/sum(job_number),
    mean_wage=label_dollar(prefix = 'R$ ')(mean_wage),
    job_number=label_number(big.mark = ',')(job_number),
    job_share=label_percent(accuracy=.01)(job_share)
  ) %>%
  select(group, mean_wage, job_number, job_share) %>%
  kable(
    col.names = c(
      'Skill group',
      'Mean hourly wage',
      'Number of emp. relationships',
      'Share of emp. relationships'
    ),
    caption = "Summary of wages and employment relationships, 2019",
  align = "llrr"
  )
```

```{r}
chart1data <- data %>%
  filter(!is.na(group)) %>%
  select(ano, mean_wage, job_number) %>%
  group_by(ano) %>%
  summarise_all(mean) %>%
  mutate(
    mean_wage_index = mean_wage/first(mean_wage),
    job_number_index = job_number/first(job_number)
  ) %>%
  ungroup()

chart1data
```

```{r}
fig41<- chart1data %>%
  ggplot( aes(x=ano, y=mean_wage_index, group=1)) +
    geom_line() +
    # ggtitle("Mean wages") +
		theme_stata() +
    theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
          axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
          legend.title=element_blank(),legend.background = element_rect(colour='white'))+
    guides(fill=FALSE) +
    ylab("Mean hourly wages (2006=1)") +
    xlab(NULL)

ggsave(str_interp("${fig_dir}/fig-4-1.png"),width=8,height=5)

fig41
```

```{r,fig.width = 8}
fig43 <- chart1data %>%
  ggplot( aes(x=ano, y=job_number_index, group=1)) +
    geom_line() +
    # ggtitle("Employment relationships") +
		theme_stata() +
    theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
          axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
          legend.title=element_blank(),legend.background = element_rect(colour='white'))+
    guides(fill=FALSE) +
    ylab("Number of employment relationships (2006=1)") +
    xlab(NULL)

ggsave(str_interp("${fig_dir}/fig-4-3.png"),width=8,height=5)

fig43
```

```{r}
chart2data <- data %>%
  select(group, ano, mean_wage, job_number) %>%
  group_by(group, ano) %>%
  summarise_all(mean) %>%
  mutate(
    mean_wage_index = mean_wage/first(mean_wage),
    job_number_index = job_number/first(job_number)
  ) %>%
  ungroup() %>%
  group_by(group) %>%
  mutate(
    mean_wage_2006 = sum(mean_wage*as.integer(ano==2006))
  ) %>%
  ungroup() %>%
  arrange(mean_wage_2006, ano) %>%
  mutate(
    order=seq(1, n())
  )

chart2data
```

```{r}
fig42 <- chart2data %>%
  filter(!is.na(group)) %>%
  ggplot( aes(x=ano, y=mean_wage_index, group=group, color=reorder(group, order))) +
    geom_line(size=1) +
    scale_color_manual(values=c('#d7191c','#2c7bb6','#4de678','#fdae61'))+
    # ggtitle("Mean wages by skill group") +
		theme_stata() +
    theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
          axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
          legend.title=element_blank(),legend.background = element_rect(colour='white'))+
												guides(fill=FALSE) +
    ylab("Mean hourly wages (2006=1)") +
    xlab(NULL)

ggsave(str_interp("${fig_dir}/fig-4-2.png"),width=8,height=5)

fig42
```

```{r}
fig44 <- chart2data %>%
  filter(!is.na(group)) %>%
  ggplot(
    aes(x=ano, y=job_number_index, group=group, color=reorder(group, order))
  ) +
    geom_line(size=1) +
    scale_color_manual(values=c('#d7191c','#2c7bb6','#4de678','#fdae61'))+
    # scale_color_viridis(discrete = TRUE) +
    # ggtitle("Employment relationships") +
		theme_stata() +
    theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
      axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
      legend.title=element_blank(),legend.background = element_rect(colour='white'))+
    guides(fill=FALSE) +
    ylab("Number of employment relationships (2006=1)") +
    xlab(NULL)


ggsave(str_interp("${fig_dir}/fig-4-4.png"),width=8,height=5)

fig44
```

```{r}
chart3data <- data %>%
  filter(ano %in% cbind(seq(2006, 2018, 3), 2019)) %>%
  filter(!is.na(group)) %>%
  group_by(group) %>%
  arrange(ano, .by_group = TRUE) %>%
  mutate(
    delta_job_number = case_when(
      ano == 2019 ~ job_number/first(job_number) - 1,
      .default = job_number/lag(job_number) - 1
    ),
    interval = case_when(
      ano == 2019 ~ paste(first(ano), ano, sep = '-'),
      .default = paste(lag(ano), ano, sep = '-')
    ),
    mean_wage_2006 = sum(mean_wage*as.integer(ano==2006))
  ) %>%
  ungroup() %>%
  filter(ano != 2006) %>%
  arrange(mean_wage_2006, ano) %>%
  select(interval, group, delta_job_number, mean_wage_2006) %>%
  mutate(
    order=seq(1, n())
  )

chart3data
```

```{r warning=FALSE}
fig45 <- chart3data %>% ggplot(aes(x=reorder(group, order),y=delta_job_number,fill=group, alpha=reorder(interval, order)))+
											geom_bar(stat='identity',position='dodge')+
                      theme_stata()+
											labs(x='',y='Change (% pts)')+
											scale_fill_manual(values=c('#d7191c','#fdae61','#4de678','#2c7bb6'))+
											theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
												axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
												legend.title=element_blank(),legend.background = element_rect(colour='white'))+
												guides(fill=FALSE)

ggsave(str_interp("${fig_dir}/fig-4-5.png"),width=8,height=5)

fig45
```

## Export data

```{r}

file <- str_interp('${datadir}/WITS/WITS_trades.parquet')
wits <- tbl(con, file) %>%
  filter(year < 2020) %>%
  collect()

dbDisconnect(con, shutdown=TRUE)

wits %>%
  head()
```

```{r}
fig46 <- wits %>%
  mutate (
    trade_value = `Trade Value 1000USD`/(10^6)
  ) %>%
  ggplot(
    aes(x=year, y=trade_value, fill=reporter)
  ) +
    geom_area() +
    scale_fill_manual(values=c('#d7191c','#2c7bb6','#4de678','#fdae61', '#0ddead'))+
    scale_x_continuous(breaks = seq(2006,2019)) +
    # ggtitle("Employment relationships") +
		theme_stata() +
    theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
      axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
      legend.title=element_blank(),legend.background = element_rect(colour='white'))+
    ylab("Trade Value (billions of US dollars)") +
    xlab(NULL)


ggsave(str_interp("${fig_dir}/fig-4-6.png"),width=8,height=5)

fig46
```

```{r}
fig47 <- wits %>%
  mutate (
    quantity = Quantity/10^9
  ) %>%
  ggplot(
    aes(x=year, y=quantity, fill=reporter)
  ) +
    geom_area() +
    scale_fill_manual(values=c('#d7191c','#2c7bb6','#4de678','#fdae61', '#0ddead'))+
    scale_x_continuous(breaks = seq(2006,2019)) +
    # ggtitle("Employment relationships") +
		theme_stata() +
    theme(axis.ticks.x = element_blank(), plot.background = element_rect(fill='white'),
      axis.text.x=element_text(angle=-45,size=8,hjust=0),axis.text.y=element_text(angle=0),
      legend.title=element_blank(),legend.background = element_rect(colour='white'))+
    ylab("Quantity (billions of items)") +
    xlab(NULL)


ggsave(str_interp("${fig_dir}/fig-4-7.png"),width=8,height=5)

fig47
```

# References
