==> 05_CNAE.Rmd <==
---
title: "Computers and jobs - CNAE"
output: 
  rmarkdown::github_document:
    df_print: kable
bibliography: ref-mono.bib
csl: associacao-brasileira-de-normas-tecnicas-ipea.csl
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook for the exploratory analysis and preparation of data from the Classificação Nacional das Atividades Econômicas (CNAE), managed by Instituto Brasileiro de Geografia e Estatística (IBGE), as made available by [Base dos Dados](https://basedosdados.org/).

```{r}
setwd("D:/OneDrive/R Workspace/computers_and_jobs")
# install.packages(\"dotenv\")
library("dotenv")
# install.packages('arrow')
library(arrow, warn.conflicts = FALSE)
# install.packages("tidyverse")
library("tidyverse")
```

First, I need to setup the GCP project data for using the `basedosdados` package to get the data.

```{r}
# Defina o seu projeto no Google Cloud
load_dot_env()
project_id <- Sys.getenv("GCP_PROJECT_ID")
basedosdados::set_billing_id(project_id)
bigrquery::bq_auth(email = Sys.getenv("GCP_EMAIL"))
```

```{r}
peek <- basedosdados::read_sql("SELECT * FROM `basedosdados.br_bd_diretorios_brasil.cnae_2` LIMIT 1")
peek %>% dplyr::glimpse()
```

Now, I download the CNAE data from basedosdados.

```{r}

sql <- "
SELECT
  cnae.cnae_2,
  cnae.descricao,
  cnae.grupo,
  cnae.descricao_grupo,
  cnae.divisao,
  cnae.descricao_divisao,
  cnae.secao,
  cnae.descricao_secao
FROM
  basedosdados.br_bd_diretorios_brasil.cnae_2 AS cnae
"

cnae <- basedosdados::read_sql(sql)
cnae %>% head()
```

For some unknown reason, there was some problem with the value of `descricao_secao` just for one section. I'll update the description of that section before saving.

```{r}
cnae %>%
  filter(secao=='E') %>%
  select(descricao_secao) %>%
  head()
```

```{r}
cnae <- cnae %>%
  mutate(
    descricao_secao = case_when(
      secao=='E' ~ 'Água, esgoto, atividades de gestão de resíduos e descontaminação',
      .default=descricao_secao
    )
  )

cnae %>%
  filter(secao=='E') %>%
  select(descricao_secao) %>%
  head()
```

And, just to be sure, check some other section

```{r}
cnae %>%
  filter(secao=='F') %>%
  select(descricao_secao) %>%
  head()
```

Finally, I save the data to file.

```{r}
data_dir <- "data/CNAE/"
file <- str_interp("${data_dir}CNAE.parquet")

write_parquet(cnae, file)
```
