==> 07_RAIS_CNAE_CBO_IPCA.Rmd <==
---
title: "Computers and jobs - Merging of RAIS, CNAE and CBO"
output:
  rmarkdown::github_document:
    df_print: kable
bibliography: ref-mono.bib
csl: associacao-brasileira-de-normas-tecnicas-ipea.csl
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook for the preparation and merge of data from the Relação Anual de Informações Sociais (RAIS), Classificação Nacional das Atividades Econômicas (CNAE), Cadastro Brasileiro de Ocupações (CBO) and Índice Nacional de Preços ao Consumidor Amplo (IPCA).

```{r}
setwd("D:/OneDrive/R Workspace/computers_and_jobs")
# install.packages(\"dotenv\")
library("dotenv")
# install.packages('arrow')
library(arrow, warn.conflicts = FALSE)
# install.packages("tidyverse")
library("tidyverse")
library("duckdb")
```

# Setting up file paths and DuckDB

```{r}
load_dot_env()
base_datadir <- Sys.getenv("DATA_DIR")
rais_datadir <- str_interp("${base_datadir}RAIS/")
```

As the RAIS data is too big to fit in memory at one time and is partitioned by UF and year, I'll use `dbplyr` and `duckdb` to handle the loading of the files for me.

First, I'll setup the connection.

```{r}
temp_db_dir <- Sys.getenv("DUCK_DB_DIR")
temp_db_mem <- Sys.getenv("DUCK_DB_MEM")

con <- dbConnect(duckdb(), dbdir = temp_db_dir)
dbExecute(conn = con, str_interp("PRAGMA memory_limit='${temp_db_mem}'"))
```

# Setting up tables

Now, I'll setup the RAIS dataset and read only the first rows just to see columns.

```{r}
rais_filepath <- str_interp("'${rais_datadir}RAIS_*_*.parquet'")

rais <- tbl(con, rais_filepath)

rais %>%
  head() %>%
  collect()
```

And the same for CBO, CNAE and IPCA datasets.

```{r}
cbo <- tbl(con, str_interp("'${base_datadir}CBO.parquet'"))

cbo %>%
  head() %>%
  collect()
```

```{r}
cnae <- tbl(con, str_interp("'${base_datadir}CNAE.parquet'"))
cnae %>%
  head() %>%
  collect()
```

```{r}
ipca <- tbl(con, str_interp("'${base_datadir}IPCA.parquet'"))
ipca %>%
  head() %>%
  collect()
```

# Merging the datasets

And merge the CBO and CNAE datasets.

```{r}
rais <- tbl(con, rais_filepath) %>%
  left_join(
    cbo %>% select(cbo_2002, group),
    by=join_by(cbo_2002==cbo_2002)
  ) %>%
  left_join(
    cnae %>% select(cnae_2, secao, descricao_secao),
    by=join_by(cnae_2==cnae_2)
  ) %>%
  left_join(
    ipca,
    by=join_by(ano==ano)
  )

rais %>%
  head() %>%
  collect()
```

# Correcting the wage values

Now, I correct the wage values to 2021 prices using IPCA and convert to hourly wages (so that the different contract hours are take into account).

```{r}
rais <- rais %>%
  mutate(
    mean_wage = valor_remuneracao_media*multiplier_2021,
    mean_wage = mean_wage/(quantidade_horas_contratadas*4)
  )

rais %>%
  head() %>%
  collect()
```

# Aggregation the database

Finally, I aggregate the data by the fixed effects and possible cofactors.

```{r}
rais <- rais %>%
      group_by(
        ano,
        sigla_uf,
        secao,
        descricao_secao,
        group,
        grau_instrucao_apos_2005,
        sexo,
        raca_cor,
        idade
      ) %>%
      summarise(
        valor_remuneracao_media=mean(valor_remuneracao_media),
        valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
        mean_wage=mean(mean_wage),
        job_number=n()
      ) %>%
      ungroup()
```

```{r}
tictoc::tic()
rais_agg <- rais %>%
  collect()

tictoc::toc()
```

```{r}
rais_agg %>% head()
```


# Saving the results to disk

```{r}
write_parquet(rais_agg, str_interp("${base_datadir}/RAIS_AGG.parquet"))
```
