==> 09_regressions.Rmd <==
---
title: "Computers and jobs - Regressions"
output:
  rmarkdown::github_document:
    toc: true
    toc_depth: 2
    number_sections: true
    df_print: kable
bibliography: ref-mono.bib
csl: associacao-brasileira-de-normas-tecnicas-ipea.csl
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook for the first set of regressions developed for the Undergrad thesis. As always, the first part is for the setup of the libraries and paths.

# Initial setup

```{r}
setwd("D:/OneDrive/R Workspace/computers_and_jobs")
# install.packages(\"dotenv\")
library("dotenv")
# install.packages('arrow')
library(arrow, warn.conflicts = FALSE)
# install.packages("tidyverse")
library("tidyverse")
library("duckdb")
if(!require(modelsummary)){install.packages('modelsummary');require(modelsummary)}
library("gtsummary")
if(!require(plm)){install.packages("plm");require(plm)}
```

```{r}
load_dot_env()
datadir <- Sys.getenv("DATA_DIR")
file <- str_interp('${datadir}Final_db.parquet')
```

At first, I'm not using the additional covariates present in the database (gender, race, age and school level) as they don't exactly fit with the TWFE approach documented previously. So, I'll use DuckDB to load only a subset of the columns and execute an initial grouping.

```{r}
temp_db_dir <- Sys.getenv("DUCK_DB_DIR")
temp_db_mem <- Sys.getenv("DUCK_DB_MEM")

con <- dbConnect(duckdb(), dbdir = temp_db_dir)
dbExecute(conn = con, str_interp("PRAGMA memory_limit='${temp_db_mem}'"))
```

```{r}
tictoc::tic()

data <- tbl(con, file) %>%
  filter(
    !is.na(secao) & !is.na(ano) & !is.na(group)
  ) %>%
  filter (
    ano < 2020
  ) %>%
  group_by(
    descricao_secao,
    ano,
    group
  ) %>%
  summarise(
    exposure=mean(exposure),
    valor_remuneracao_media=mean(valor_remuneracao_media),
    valor_remuneracao_media_sm=mean(valor_remuneracao_media_sm),
    mean_wage=mean(mean_wage),
    job_number=sum(job_number)
  ) %>%
  ungroup %>%
  mutate(
    log_wage = log(mean_wage + 1)
  ) %>%
  collect()

dbDisconnect(con, shutdown=TRUE)

data <- data %>%
  rename(
    secao = descricao_secao
  ) %>%
  mutate(
    secao = as.factor(secao),
    group = as.factor(group) %>% relevel('NRM'),
    ano = as.factor(ano),
    id = factor(paste(secao, group, sep="-"))
  )

tictoc::toc()
data %>% head()
```

Also, in order to support some of the models, I need to add the share of employment relationships of each group in respect of industry-year.

```{r}
data2 <- data %>%
  group_by(ano) %>%
  mutate(
    job_share = case_when(
      ano == 2006 ~ job_number/ sum(job_number),
      .default = 0
    )
  ) %>%
  ungroup() %>%
  group_by(id) %>%
  mutate(
    job_share = order_by(ano, cumsum(job_share)),
    log_job = log(job_number)
  ) %>%
  ungroup()

data2 %>% head()
```

```{r include=FALSE}

default_table <- function(models, coefs, robust='stata', ...) {
  tb <- modelsummary(
    models,
    coef_rename = coefs,
    gof_omit = 'AIC|BIC|RMSE',
    estimate = '{estimate}{stars} ({std.error})',
    stars=TRUE,
    statistic = "(p $\\approx$ {p.value})",
    vcov=robust,
    ...
  )

  return(tb)
}

gtsummary_table <- function(model, include, label, robust=NULL, ...) {
  
  tb <- tbl_regression(
    model,
    include = include,
    label = label,
    intercept=TRUE,
    estimate_fun = function(x) style_number(x, digits = 4),
    ...
    ) 
  if(!is.null(robust)) {
    if (robust!=FALSE) {
      tb <- tbl_regression(
        model,
        include = include,
        label = label,
        intercept=TRUE,
        estimate_fun = function(x) style_number(x, digits = 4),
        tidy_fun = partial(tidy_robust, vcov = robust, quiet = TRUE),
        ...
        )
    }
  }

  tb <- tb %>%
  bold_labels() %>%
  add_significance_stars(
    hide_se = TRUE,
    hide_p=FALSE,
    pattern = '{estimate}{stars} ({std.error})'
  ) %>%
  modify_header(
    estimate ~ '**Coef. (SE)**',
    p.value ~ '**p-value**'
    ) %>%
  modify_footnote(estimate ~ "SE = Standard error", abbreviation = TRUE) %>%
  add_glance_table(
    include=c(
      'r.squared',
      'adj.r.squared',
      'nobs'
    )
  )
  return(tb)
}
```

# Effectiveness of the task-based model for the Brazilian formal labor market

Lastly, although initially as fixed effects to evaluate the influence of exposure to computers on the mean_wage inequality, it could be useful to evaluate the influence of the skill composition of the occupations with the mean_wage inequality. It is useful as allows an evaluation of the capability of task model itself to explain (at least, part of) the mean_wage inequality on the formal jobs in Brazil.

To do that, I start from the following model:

$$\log(y_{j, c, t}) = \beta Exposure_{j, t} + \alpha_j + \lambda_c + \theta_t + \varepsilon_{j, c, t},$$

where $Exposure_{j, t}$ is the industry-year measure of exposure to computers, $\alpha_j$ is an industry fixed effect, $\lambda_c$ is a skill classification fixed effect, and $\theta_t$ is a period fixed effect. Thus, $\beta$ will identify the effects of the changes in exposure to computers within each outcome. The outcome here can be either the log wage, where $y_j, c, t$ stands for the mean wage for occupations in the industry $j$ and category $c$ on year $t$, or the log of the number of employment relationships, where $y_j, c, t$ stands for the sum of employment relationships of occupations in the industry $j$ and category $c$ on year $t$.

Then, I estimate three derived models based on the One-Way Fixed Effects Model described in @baltagi2021 [p. 15-19] on the time variable and include the skill groups of the occupations as explainable variables. I'll evaluate one specification with the skill groups as the only explainable variable, and another specification with industry sector as explainable variables, to check for omitted variable bias.

## Only skill groups as explainable variables

So, this model becomes:

$$\left[\log(y_{j, c, t}) - \overline{\log(y_{t})}\right] = (\lambda_c - \overline{\lambda}) + (\varepsilon_{j, c, t} - \overline{\varepsilon_{t}})$$

where $\overline{\log(y_{t})}$ and $\overline{\lambda}$ are the time demeaned versions of the outcome variables and skill classification group dummies.

```{r}
fe_reg_w_1 <- plm(log_wage ~ group,
             data = data2,
             model = 'within',
             effect = 'time',
             index = c('id', 'ano'))

fe_reg_j_1 <- plm(log_job ~ group,
             data = data2,
             model = 'within',
             effect = 'time',
             index = c('id', 'ano'))
```

## Skill groups and industry sectors as explainable variables

This one becomes:

$$\left[\log(y_{j, c, t}) - \overline{\log(y_{t})}\right] = (\lambda_c - \overline{\lambda}) + (\alpha_j - \overline{\alpha}) + (\varepsilon_{j, c, t} - \overline{\varepsilon_{t}})$$

where $\overline{\log(y_{t})}$, $\overline{\lambda}$, and $\overline{\alpha}$ are the time demeaned versions of the outcome variables, skill classification group dummies, and industry sector dummies.

```{r}
fe_reg_w_2 <- plm(log_wage ~ group + secao,
             data = data2,
             model = 'within',
             effect = 'time',
             index = c('id', 'ano'))

fe_reg_j_2 <- plm(log_job ~ group + secao,
             data = data2,
             model = 'within',
             effect = 'time',
             index = c('id', 'ano'))
```

## Results of the skill groups specifications

```{r}
tb_fe_reg_w_1 <- gtsummary_table(
  fe_reg_w_1,
  include=c('group'),
  label=list(
    'group'='Skill group'
  ),
  robust='HC1'
)
tb_fe_reg_j_1 <- gtsummary_table(
  fe_reg_j_1,
  include=c('group'),
  label=list(
    'group'='Skill group'
  ),
  robust='HC1'
)

tb_fe_reg_w_2 <- gtsummary_table(
  fe_reg_w_2,
  include=c('group'),
  label=list(
    'group'='Skill group'
  ),
  robust='HC1'
)
tb_fe_reg_j_2 <- gtsummary_table(
  fe_reg_j_2,
  include=c('group'),
  label=list(
    'group'='Skill group'
  ),
  robust='HC1'
)

tbl_merge(
  list(tb_fe_reg_w_1, tb_fe_reg_w_2, tb_fe_reg_j_1, tb_fe_reg_j_2),
  c('**log(wage)**', '**log(wage)**', '**log(job)**', '**log(job)**')
)
```

Now, more interesting results. Although the main focus of the thesis was to investigate the relation of computers and wage polarization, the approach would only be valid if the task model itself could be considered able to explain the wage inequality in the Brazilian formal market.

And that's what we can conclude from the models above. Even without the industry sector fixed effects, the model get a $R^2\approx0.525$ and a $\bar{R^2}\approx0.519$, which means that the main composition of the tasks of the occupations alone can explain more than half of the variability of the wages in the formal market. Of course, this results can be biased by omitted variables, like school level or signalization of diplomas, but the results are way better than the expected. With the inclusion of industry sector fixed effects, the $R^2$ and $\bar{R^2}$ change to $0.916$ and $0.913$, respectively, but the increase is mixed with the explanation capability of the industry sectors themselves, which is not the focus of this thesis.

Now, the most important results: all of the variables are extremely significant and the inclusion of the industry sector effects didn't change the coefficients, which is a good indicative of non existence of omitted variable bias. Again, further specifications could be thought to investigate this bias, but that would need additional theoretical investigation to compose the model.

The interpretation of the coefficients in this case, thankfully, is very simple: a worker with an occupation composed mostly of tasks belonging to the RM group is expected to receive, on average, $14,7\%$ more for his hourly wage than a worker with an occupation composed mostly of tasks belonging to the NRM group. For the NRC and RC groups, the premium is $103,1\%$ and $39,2\%$, respectively. However, if we consider the order of the groups to be, from lowest to highest, { NRM, RM, RC, NRC }, there is no wage ***polarization*** effect. Although the task model exhibits an excellent power of explanation of the Brazilian formal labor market, the values encountered for the coefficients indicates a monotone change of the wages.

## Industry sectors as the only explainable variables

Just to dig a little bit more about omitted variable bias in the industry sector fixed effects, I'll regress wages on industry sector only and compare with the fixed effects in the previous model with industry sector and skill groups. So, this model becomes:

$$\left[\log(y_{j, c, t}) - \overline{\log(y_{t})}\right] = (\alpha_j - \overline{\alpha}) + (\varepsilon_{j, c, t} - \overline{\varepsilon_{t}})$$

where $\overline{\log(y_{t})}$ and $\overline{\alpha}$ are the time demeaned versions of the outcome variables and industry sector dummies.

```{r}
fe_reg_w_3 <- plm(log_wage ~ secao,
             data = data2,
             model = 'within',
             effect = 'time',
             index = c('id', 'ano'))

fe_reg_j_3 <- plm(log_job ~ secao,
             data = data2,
             model = 'within',
             effect = 'time',
             index = c('id', 'ano'))
```

## Results of the industry sector specifications

```{r}
tb2_fe_reg_w_2 <- gtsummary_table(
  fe_reg_w_2,
  include=c('secao'),
  label=list(
    'secao'='CNAE industry sector'
  ),
  robust='HC1'
)
tb_fe_reg_w_3 <- gtsummary_table(
  fe_reg_w_3,
  include=c('secao'),
  label=list(
    'secao'='CNAE industry sector'
  ),
  robust='HC1'
)
tb2_fe_reg_j_2 <- gtsummary_table(
  fe_reg_j_2,
  include=c('secao'),
  label=list(
    'secao'='CNAE industry sector'
  ),
  robust='HC1'
)
tb_fe_reg_j_3 <- gtsummary_table(
  fe_reg_j_3,
  include=c('secao'),
  label=list(
    'secao'='CNAE industry sector'
  ),
  robust='HC1'
)
tbl_merge(
  list(tb2_fe_reg_w_2, tb_fe_reg_w_3, tb2_fe_reg_j_2, tb_fe_reg_j_3),
  c('**log(wage)**', '**log(wage)**', '**log(job)**', '**log(job)**')
)
```

The main interest in this results is the possibility of bias in the industry sector coefficients, which didn't appear, as all the coefficients kept the same with and without the skill group variable.

# Relation of computer adoption and the labor market

Here, I also start from the following model:

$$\log(y_{j, c, t}) = \beta Exposure_{j, t} + \alpha_j + \lambda_c + \theta_t + \varepsilon_{j, c, t},$$

where $Exposure_{j, t}$ is the industry-year measure of exposure to computers, $\alpha_j$ is an industry fixed effect, $\lambda_c$ is a skill classification fixed effect, and $\theta_t$ is a period fixed effect. Thus, $\beta$ will identify the effects of the changes in exposure to computers within each outcome. The outcome here can be either the log wage, where $y_j, c, t$ stands for the mean wage for occupations in the industry $j$ and category $c$ on year $t$, or the log of the number of employment relationships, where $y_j, c, t$ stands for the sum of employment relationships of occupations in the industry $j$ and category $c$ on year $t$.

Then, I estimate three derived models based on the Two Way Fixed Effects Model described in @baltagi2021. I'll detail each model right before its estimation.

## First specification - reproduction of @nberw30400

First, I'll try to reproduce the same analysis of @nberw30400, without considering the predominant task set of the occupations. For that, the first thing I need to do is aggregate the data.

```{r}
nber30400 <- data%>%
  filter(!is.na(secao) & !is.na(ano)) %>%
  group_by(secao, ano) %>%
  summarise(
    mean_wage = mean(mean_wage),
    exposure = mean(exposure),
    job_number = sum(job_number)
  ) %>%
  mutate(
    log_wage = log(mean_wage + 1),
    log_job = log(job_number)
  ) %>%
  ungroup() %>%
  group_by(ano) %>%
  mutate(
    job_share = case_when(
      ano == 2006 ~ job_number/ sum(job_number),
      .default = 0
    )
  ) %>%
  ungroup() %>%
  group_by(secao) %>%
  mutate(
    job_share = cumsum(job_share)
  ) %>%
  ungroup()

nber30400 %>% head()
```

Now, I regress exposure on log_remuneracao_media with plm. This package provides a series of estimators for panel data, including the Two-Way Fixed Effects described in @baltagi2021 [p. 47-49], which is used here. So, the estimated model has the following form:

$$\widetilde{\log(y)}_{j,t} = \beta\widetilde{Exp}_{j,t} + \widetilde{\varepsilon}_{j,t}$$

where $\widetilde{\log(y)}_{j,t}$ and $\widetilde{Exp}_{j,t}$ are the double demeaned versions of the outcome variables (log wages or log of employment relationships) and exposure for computers measure. Also, as this model doesn't consider the skill groups of occupations, the outcome is indexed only by industry-year.

```{r}
twfe_reg_w <- plm(log_wage ~ exposure,
             data = nber30400,
             model = 'within',
             effect = 'twoways',
             index = c('secao','ano'))
```

```{r}
twfe_reg_j <- plm(log_job ~ exposure,
             data = nber30400,
             model = 'within',
             effect = 'twoways',
             index = c('secao','ano'))
```

```{r}
tb_twfe_reg_w <- gtsummary_table(
  twfe_reg_w,
  include = c('exposure'),
  label = list('exposure'='Exposure to computers'),
  robust = 'HC1'
)

tb_twfe_reg_j <- gtsummary_table(
  twfe_reg_j,
  include=c('exposure'),
  label = list('exposure'='Exposure to computers'),
  robust = 'HC1'
)

tbl_merge(
  list(tb_twfe_reg_w, tb_twfe_reg_j),
  c('**log(wages)**', '**log(jobs)**')
)
```

The results show that the `Exposure to computers` measure alone does not explain any variation in the log mean_wage of the brazilian formal market, both by the non significant result for the variable itself and by the $R^2$ for the within estimator.

## Second specification - adding occupation classification

As a second step, I'll modify the model to incorporate the task compositions of the occupations through the discrete classification of the predominant skill group of each occupation. With the data aggregated by industry sector, year and task group, I regress the log_wages on the exposure measure.In this case, the estimated model is:

$$\widetilde{\log(y)}_{j, c, t} = \gamma\widetilde{Exp}_{j,t} + \widetilde{\varepsilon}_{j, c, t}$$

where, again, $\widetilde{\log(y)}_{j, c, t}$ and $\widetilde{Exp}_{j,t}$ are the double demeaned versions of the outcome variables (log wages or log of employment relationships) and exposure for computers measure.

```{r}
twfe_reg_w_2 <- plm(log_wage ~ exposure,
             data = data2,
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
twfe_reg_j_2 <- plm(log_job ~ exposure,
             data = data2,
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
tb_twfe_reg_w_2 <- gtsummary_table(
  twfe_reg_w_2,
  include = c('exposure'),
  label = list('exposure'='Exposure to computers'),
  robust = 'HC1'
)

tb_twfe_reg_j_2 <- gtsummary_table(
  twfe_reg_j_2,
  include = c('exposure'),
  label = list('exposure'='Exposure to computers'),
  robust='HC1'
)

tbl_merge(
  list(tb_twfe_reg_w_2, tb_twfe_reg_j_2),
  c('**log(wages)**', '**log(jobs)**')
)
```

Again, the results show that the `Exposure to computers` measure alone does not explain any variation in the log mean_wage of the brazilian formal market, both by the non significant result for the variable itself and by the $R^2$ for the within estimator. As the result of both these specifications indicate a non significant value for the coefficient, at least it seems to be consistent.

## Third specification - Interacting group with exposure

Two last test using the @baltagi2021 TWFE will be made by interacting the predominant skill group of the occupations with the exposure to computers measure, so we can investigate if the exposure to computer has heterogeneous effects through the different skill groups. To prevent the correlation between the exposure to computers measure and the RC occupations of the first period, used to construct the measure, we exclude the first period data in this evaluation. So, the two following models are estimated here:

$$\widetilde{\log(y)}_{j, c, t} = \delta\widetilde{Exp\times\lambda}_{j,c,t} + \widetilde{\varepsilon}_{j, c, t}$$

and

$$\widetilde{\log(y)}_{j, c, t} = \rho\widetilde{Exp}_{j,t} + \delta\widetilde{Exp\times\lambda}_{j,c,t} + \widetilde{\varepsilon}_{j, c, t}$$

where, again, $\widetilde{\log(y)}_{j, c, t}$ and $\widetilde{Exp}_{j,t}$ are the double demeaned versions of the outcome variables (log wages or log of employment relationships) and exposure for computers measure, and $\delta\widetilde{Exp\times\lambda}_{j,c,t}$ is the double demeaned version of the multiplication of exposure for computers measure and skill group classification vector.

```{r}
twfe_reg_w_3 <- plm(log_wage ~ exposure:group,
             data = data2 %>% filter(ano != '2006'),
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
twfe_reg_j_3 <- plm(log_job ~ exposure:group,
             data = data2 %>% filter(ano != '2006'),
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
twfe_reg_w_4 <- plm(log_wage ~ exposure + exposure:group,
             data = data2 %>% filter(ano != '2006'),
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
twfe_reg_j_4 <- plm(log_job ~ exposure + exposure:group,
             data = data2 %>% filter(ano != '2006'),
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
tb_twfe_reg_w_3 <- gtsummary_table(
  twfe_reg_w_3,
  include = c('exposure:group'),
  label = list(
    'exposure:group'='Exposure to computers x Skill type'
    ),
  robust = 'HC1'
)

tb_twfe_reg_w_4 <- gtsummary_table(
  twfe_reg_w_4,
  include = c('exposure', 'exposure:group'),
  label = list(
    'exposure:group'='Exposure to computers x Skill type'
    ),
  robust = 'HC1'
)

tb_twfe_reg_j_3 <- gtsummary_table(
  twfe_reg_j_3,
  include = c('exposure:group'),
  label = list(
    'exposure:group'='Exposure to computers x Skill type'
    ),
  robust = 'HC1'
)

tb_twfe_reg_j_4 <- gtsummary_table(
  twfe_reg_j_4,
  include = c('exposure', 'exposure:group'),
  label = list(
    'exposure:group'='Exposure to computers x Skill type'
    ),
  robust = 'HC1'
)

tbl_merge(
  list(tb_twfe_reg_w_4, tb_twfe_reg_w_3, tb_twfe_reg_j_4, tb_twfe_reg_j_3),
  c('**log(wage)**', '**log(wage)**', '**log(job)**', '**log(job)**')
)
```

Now, some interesting results. On **(1)**, with just the interactions as explainable variables, none of the coefficients are significant at 5% level, but the *p-value* of the interaction between the measure and the RC occupations is $\approx0.054$, which seems good enough to deserve some consideration. But, contrary to my hypothesis, in this case the exposure to computers would be responsible to an increase in the mean wage of this group of occupations. As the measure itself is not so clear to explain, the better approach of interpretation I can think is to describe what would happen in the wage based on the means and standard errors of both of the variables (explained and explainable). I'll not evolve on this now, as I didn't do any descriptive statistics work, but I'll return on this later.

On **(2)**, the things change a little. As the NRM interaction is dropped, all of the other coefficients get bigger absolute values. But the variance of the estimators don't change too much, which causes the *p-value* of all of the coefficients to decrease. So, in this specification, we get the NRC and RC to be significant at 5% level, but it seems more like a mechanical (and accidental) p-hack. Before finish this evaluation, I'll try a third variation here, setting the NRC interaction (the closest to 0) to be dropped.

```{r}
data3 <- data2 %>%
  mutate(
    group = as.factor(group) %>% relevel('NRC'),
  ) %>%
  filter(ano != '2006')

twfe_reg_w_5 <- plm(log_wage ~ exposure + exposure:group,
             data = data3,
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))

twfe_reg_j_5 <- plm(log_job ~ exposure + exposure:group,
             data = data3,
             model = 'within',
             effect = 'twoways',
             index = c('id','ano'))
```

```{r}
tb_twfe_reg_w_5 <- gtsummary_table(
  twfe_reg_w_5,
  include=c('exposure', 'exposure:group'),
  label = list('exposure'='exposure',
               'exposure:group'='Exposure to computers x Skill type'),
  robust = 'HC1'
)

tb_twfe_reg_j_5 <- gtsummary_table(
  twfe_reg_j_5,
  include=c('exposure', 'exposure:group'),
  label = list('exposure'='exposure',
               'exposure:group'='Exposure to computers x Skill type'),
  robust = 'HC1'
)

tbl_merge(
  list(tb_twfe_reg_w_5, tb_twfe_reg_j_5),
  c('**log(wage)**', '**log(job)**')
)
```

As expected, in **(3)** the significance of the RC interaction doesn't keep significant (by far, with a $0.342$ *p-value*) with the change of group to be the ground level. Now, the RM and NRM interactions (which have signs opposed to NRC interaction) show a 5% level significance. But the results from **(2)** and **(3)** doesn't appear to be robust enough to be reported, so I intend to keep just the **(1)** specification on the final version of the thesis.

# References
